1. env_pre.mk: Add additional build-tree states and unify them under the $(BUILD_TREE_STATE) variable. 2. Makefile: Resolve issue spotted by Mitani-san because it was an overlooked corner case that wasn't properly expressed until $(BUILD_TREE_STATE) was introduced. 3. testscripts/... : resolve some issues with the test scripts.

+++ testscripts/build/build_test_function.sh
@@ -30,9 +30,14 @@ setup_env() {
 		tst_brkm tst_exit TBROK 'Failed to create tmp_builddir';
 	fi
 	if tmp_destdir=$(mktemp -d) ; then
-		vars="$vars $destdir"
+		vars="$vars $tmp_destdir"
 	else
 		tst_brkm tst_exit TBROK 'Failed to create tmp_destdir';
+	fi
+	if tmp_prefix=$(mktemp -d) ; then
+		vars="$vars $tmp_prefix"
+	else
+		tst_brkm tst_exit TBROK 'Failed to create tmp_prefix';
 	fi
 	if tmp_srcdir=$(mktemp -d) ; then
 		vars="$vars $tmp_srcdir"
@@ -40,12 +45,25 @@ setup_env() {
 		tst_brkm tst_exit TBROK 'Failed to create tmp_srcdir';
 	fi
 	trap cleanup EXIT
+	cat <<EOF
+================================
+SUMMARY OF VARIABLES:
+================================
+builddir 	-> $tmp_builddir
+destdir		-> $tmp_destdir
+prefix		-> $tmp_prefix
+srcdir		-> $tmp_srcdir
+================================
+EOF
+
 }
 
 cleanup() {
-	cd /
-	trap '' EXIT
-	rm -Rf $vars
+	if [ "x${NO_CLEAN:-0}" != 1 ] ; then
+		cd /
+		trap '' EXIT
+		rm -Rf $vars
+	fi
 }
 
 # Pull from CVS.
@@ -53,7 +71,7 @@ cvs_pull() {
 
 	export CVSROOT=:pserver:anonymous@ltp.cvs.sourceforge.net:/cvsroot/ltp
 
-	if ([ -f ~/.cvspass ] || touch ~/.cvspass) ; then
+	if ( [ -f ~/.cvspass ] || touch ~/.cvspass ) ; then
 		cvs -d$CVSROOT login && cvs -z3 export -f -r HEAD ltp && srcdir="$PWD/ltp"
 	fi
 
@@ -87,15 +105,17 @@ pull_scm() {
 # set it up to pass DESTDIR as well so it will properly set the installdir
 # global and thus I won't need to include the same checks down below...
 configure() {
+
 	abspath=$(readlink -f "$testscript_dir/../../scripts/abspath.sh")
+
 	if [ "x$2" != x ] ; then
 		(test -d "$2" || mkdir -p "$2") || return $?
 	fi
+
 	make -C "$1" autotools || return $?
-	ARGS=
-	[ "x$3" != x ] && ARGS="--prefix=$("$abspath" $3)"
-	cd "$2" && "$1/configure" $ARGS
-	install_dir=$("$abspath" "$4/$3")
+
+	cd "$2" && "$1/configure" ${3:+--prefix=$("$abspath" $3)}
+
 }
 
 # Build a source tree.
@@ -103,25 +123,37 @@ configure() {
 # 1 - source directory
 # 2 - build directory
 build() {
-	make ${2:--C "$2"} ${1:-top_srcdir=$1} all
+	make ${2:+-C "$2"} \
+	     ${1:+-f "$1/Makefile" "top_srcdir=$1"} \
+	     ${2:+"top_builddir=$2"} \
+	     all
 }
 
 # Install the binaries and scripts from a build tree.
 #
 # 1 - source directory
 # 2 - build directory
+# 3 - DESTDIR
 install_ltp() {
-	make ${2:--C "$2"} ${1:-top_srcdir=$1} install
+	make ${2:+-C "$2"} \
+	     ${1:+-f "$1/Makefile" "top_srcdir=$1"} \
+	     ${2:+"top_builddir=$2"} \
+	     ${3:+"DESTDIR=$3"} \
+	     install
 }
 
 # Run a test on the installed tree.
 #
-# 1 install directory - installed tree, e.g. $(DESTDIR)/$(prefix)
+# 1 - install directory for tree, e.g. $(DESTDIR)/$(prefix)
 test_ltp() {
+
+	[ "x${1}" != x ] && export LTPROOT="$1"
+
 	# XXX (garrcoop): I haven't tracked down the root cause for the
 	# issue, but some versions of sed combined with some terminal
 	# configurations cause sed to block waiting for EOF on certain
 	# platforms when executing runltp. Thus, we should effectively close
 	# /dev/stdin before executing runltp via execltp.
-	cd "$1" && "$1/bin/execltp" < /dev/null
+	"${1:-.}/bin/execltp" -v < /dev/null
+
 }
